+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "social-network-react",
   "version": "0.1.0",
-  "private": true,
+  "private": false,
+  "homepage": "ttps://mirenovi.github.io/cdmx-social-network-frameworks",
   "dependencies": {
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
@@ -12,6 +13,11 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build"
+  },
+  "devDependencies": {
+    "gh-pages": "^2.0.0"
   }
 }
