+++ public/index.html
@@ -26,7 +26,7 @@
   <link href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.98.0/css/materialize.min.css" rel="stylesheet">
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
   <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-  <title>Mexican Food Lovers</title>
+  <title> Mexican Food Lovers </title>
 </head>
 
 <body>
