+++ src/App.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import Login from './components/Login/Login';// componente Login
+
+class App extends Component {
+    render() {
+        return (
+            <div className="Login">
+                <Login />
+            </div>
+        );                     
+    }
+}
+
+export default App;
