+++ src/components/Login/Login.css
@@ -21,3 +21,8 @@
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
+
+@media screen and (max-width: 980px){
+  width: 100%;
+  height: 10 vh;
+}
