+++ src/components/Login/Login.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Login from './login';
+import Login from './Login';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
